Replaced the cosmicRayIdentifier with the eventTypeIdentifier. Now that events don't have a default event type, this module should be used to set the event type in order to run the modules correctly
